moved poem type def and refactored imports
