+++ web/utils/types.ts
@@ -1,0 +1,9 @@
+export type Poem = {
+    title: string,
+    subtitle:string,
+    date: string,
+    content: string,
+    archive:"true" | "false",
+    slug:string ,
+    filename?:string,
+  }
